+++ src/krm3/__init__.py
@@ -1,3 +1,3 @@
 __author__ = 'K-Tech team'
-__version__ = '2.3.0'
+__version__ = '2.4.0'
 __name__ = 'krm3'