+++ src/krm3/config/fragments/flags.py
@@ -2,5 +2,6 @@ FLAGS = {
     'TRASFERTE_ENABLED': [],
     'TIMESHEET_ENABLED': [],
     'REPORT_ENABLED': [('boolean', True)],
+    'CONTACTS_ENABLED': [('boolean', True)],
     'DDT_ENABLED': [('boolean', False)],
 }